amountFor 안에서 지역 변수 형이 int 에서 double 로 변

+++ Refactoring/src/main/java/part/one/Customer.java
@@ -50,7 +50,7 @@ public class Customer {
     }
 
     private double amountFor( Rental each) {
-        int thisAmount = 0;
+        double thisAmount = 0;
         switch (each.getMovie().getPriceCode()) {
             case Movie.REGULAR -> {
                 thisAmount += 2;
